Add files via upload
This is a slots program that utilizes a struct to hold the win results, the multiplier of the win, and a message tailored to that win condition.  Multiple functions are utilized to create the slots display, and generate the symbols. The symbols that are involved in the slot are stored in a vector.  The win condition is checked by a switch statement and an if statement. The user can dictate their balance and how much to bet.